Move "Open Project" under standard "Open File"
